Fri, Jul 8, 2022, 7:19 PM +00:00

+++ pset4/recover/recover.c
@@ -20,7 +20,7 @@ int main(int argc, char *argv[])
         return 1;
     }
 
-    BYTE *buffer = malloc(512);
+    int *buffer = malloc(512);
     int jpegcount = 0;
     FILE *img = NULL;
     char *filename = malloc(8);
